requirements.txt
fastapi
uvicorn
cryptography
pydantic
python-jose
sqlalchemy
bcrypt==4.0.1
passlib
python-multipart
httpx
starlette
numpy
pandas